♻️ changed boto service structure
