+++ server/githubsrm/apis/__init__.py
@@ -1,5 +1,5 @@
-
-from .utils import BotoService, check_token
+from .aws import BotoService
+from .utils import check_token
 from .models import Entry
 from .checks_models import EntryCheck
 from .throttle import PostThrottle
